Move rotation to routing package
