+++ controller-server/src/main/java/com/yahoo/vespa/hosted/controller/routing/rotation/RotationState.java
@@ -1,5 +1,5 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.hosted.controller.rotation;
+package com.yahoo.vespa.hosted.controller.routing.rotation;
 
 /**
  * The possible states of a global rotation.
